make cli a package
